Move files to relative path to deploy
