+++ News Explorer/src/App.jsx
@@ -1,16 +1,16 @@
 import { useState } from "react";
 import "./App.css";
 import { Routes, Route, useNavigate } from "react-router-dom";
-import Header from "../components/Header";
-import Main from "../components/Main";
-import Footer from "../components/Footer";
-import About from "../components/About";
-import LoginModal from "../components/SigninModal";
-import SignupModal from "../components/RegisterModal";
-import SavedArticles from "../components/SavedArticles";
-import MobileMenu from "../components/MobileMenu";
+import Header from "../src/components/Header";
+import Main from "../src/components/Main";
+import Footer from "../src/components/Footer";
+import About from "../src/components/About";
+import LoginModal from "../src/components/SigninModal";
+import SignupModal from "../src/components/RegisterModal";
+import SavedArticles from "../src/components/SavedArticles";
+import MobileMenu from "../src/components/MobileMenu";
 import { authorize } from "../utils/Auth";
-import SearchResults from "../components/SearchResults";
+import SearchResults from "../src/components/SearchResults";
 import { searchArticles } from "../utils/Api";
 import { CurrentUserContext } from "../context/CurrentUserContext";
 import { checkToken } from "../utils/Auth";
