+++ News Explorer/src/components/RegisterModal.jsx
